light_sdk implementation
We added the light_sdk crate and imported the necessary modules.
We marked asynchronous functions as async and added the use of await.
We made the main function run asynchronously by marking it with #[tokio::main].
We added the necessary imports for File and Command.
We removed system_instruction because it is not used.
We updated the find_compressed_account_in_transaction function in accordance with the Light SDK.